Reduce image size.

+++ entrypoint.sh
@@ -1,29 +1,33 @@
 #!/bin/sh
 set -e
 
-WP_CONFIG=/var/www/html/wp-config.php
+if [ "$1" = "/bin/sh" ]; then
+    WP_CONFIG=/var/www/html/wp-config.php
 
-if [ ! -f $WP_CONFIG ]; then
-    echo "=> Installing Wordpress ..."
+    if [ ! -f $WP_CONFIG ]; then
+        echo "=> Installing Wordpress ..."
 
-    curl -k https://api.wordpress.org/secret-key/1.1/salt/ >> index.html
+        curl -k https://api.wordpress.org/secret-key/1.1/salt/ >> index.html
 
-    # Configure Wordpress
-    sed -i -e "/define('AUTH_KEY',         'put your unique phrase here');/,/require_once(ABSPATH . 'wp-settings.php');/d" /var/www/html/wp-config-sample.php
-    cat index.html >> /var/www/html/wp-config-sample.php
-    echo -e "define('FS_METHOD', 'direct');" >> /var/www/html/wp-config-sample.php
-    cat wp-config-footer.php >> /var/www/html/wp-config-sample.php
-    rm -f index.html
-    rm -f wp-config-footer.php
+        # Configure Wordpress
+        sed -i -e "/define('AUTH_KEY',         'put your unique phrase here');/,/require_once(ABSPATH . 'wp-settings.php');/d" /var/www/html/wp-config-sample.php
+        cat index.html >> /var/www/html/wp-config-sample.php
+        echo -e "define('FS_METHOD', 'direct');" >> /var/www/html/wp-config-sample.php
+        cat wp-config-footer.php >> /var/www/html/wp-config-sample.php
+        rm -f index.html
+        rm -f wp-config-footer.php
 
-    sed -e "s/database_name_here/$DB_NAME/
-    s/username_here/$DB_USER/
-    s/password_here/$DB_PASSWORD/
-    s/localhost/$DB_HOST/" /var/www/html/wp-config-sample.php > $WP_CONFIG
-    chown www-data:www-data $WP_CONFIG
-    chmod 755 $WP_CONFIG
+        sed -e "s/database_name_here/$DB_NAME/
+        s/username_here/$DB_USER/
+        s/password_here/$DB_PASSWORD/
+        s/localhost/$DB_HOST/" /var/www/html/wp-config-sample.php > $WP_CONFIG
+        chown www-data:www-data $WP_CONFIG
+        chmod 755 $WP_CONFIG
 
-    echo "=> Done"
-else
-    echo "=> Using an existing Wordpress"
+        echo "=> Done"
+    else
+        echo "=> Using an existing Wordpress"
+    fi
 fi
+
+exec "$@"
